+++ src/processing/processNode.js
