+++ src/processing/reduceRecurentOutput.js
